Fix some tests.
