+++ crates/platform-node/Cargo.toml
@@ -23,4 +23,7 @@ serde = { version = "1.0.140", features = ["derive"] }
 tokio = { version = "1.20.0", features = ["full"] }
 
 [dev-dependencies]
+moon_cache = { path = "../cache" }
+moon_project_graph = { path = "../project-graph" }
+insta = "1.16.0"
 pretty_assertions = "1.2.1"
