+++ crates/project-graph/tests/project_graph_test.rs
@@ -1,8 +1,6 @@
 use insta::assert_snapshot;
 use moon_cache::CacheEngine;
-use moon_config::{
-    GlobalProjectConfig, NodeConfig, NodeProjectAliasFormat, WorkspaceConfig, WorkspaceProjects,
-};
+use moon_config::{GlobalProjectConfig, WorkspaceConfig, WorkspaceProjects};
 use moon_project_graph::ProjectGraph;
 use moon_utils::string_vec;
 use moon_utils::test::{create_sandbox, create_sandbox_with_git, get_fixtures_dir};
@@ -39,28 +37,6 @@ async fn get_dependents_graph() -> ProjectGraph {
             ("c".to_owned(), "c".to_owned()),
             ("d".to_owned(), "d".to_owned()),
         ])),
-        ..WorkspaceConfig::default()
-    };
-
-    ProjectGraph::create(
-        &workspace_root,
-        &workspace_config,
-        GlobalProjectConfig::default(),
-        &CacheEngine::create(&workspace_root).await.unwrap(),
-    )
-    .await
-    .unwrap()
-}
-
-async fn get_aliases_graph(node_config: NodeConfig) -> ProjectGraph {
-    let workspace_root = get_fixtures_dir("project-graph/aliases");
-    let workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Map(HashMap::from([
-            ("noLang".to_owned(), "no-lang".to_owned()),
-            ("nodeNameOnly".to_owned(), "node-name-only".to_owned()),
-            ("nodeNameScope".to_owned(), "node-name-scope".to_owned()),
-        ])),
-        node: node_config,
         ..WorkspaceConfig::default()
     };
 
@@ -209,72 +185,6 @@ mod to_dot {
         graph.load("b").unwrap();
         graph.load("c").unwrap();
         graph.load("d").unwrap();
-
-        assert_snapshot!(graph.to_dot());
-    }
-}
-
-mod aliases {
-    use super::*;
-
-    #[tokio::test]
-    async fn loads_node_aliases_name_only() {
-        let graph = get_aliases_graph(NodeConfig {
-            alias_package_names: Some(NodeProjectAliasFormat::NameOnly),
-            ..NodeConfig::default()
-        })
-        .await;
-
-        assert_eq!(
-            graph.aliases_map,
-            HashMap::from([
-                ("pkg-bar".to_owned(), "nodeNameOnly".to_owned()),
-                ("pkg-foo".to_owned(), "nodeNameScope".to_owned())
-            ])
-        );
-    }
-
-    #[tokio::test]
-    async fn loads_node_aliases_name_scopes() {
-        let graph = get_aliases_graph(NodeConfig {
-            alias_package_names: Some(NodeProjectAliasFormat::NameAndScope),
-            ..NodeConfig::default()
-        })
-        .await;
-
-        assert_eq!(
-            graph.aliases_map,
-            HashMap::from([
-                ("pkg-bar".to_owned(), "nodeNameOnly".to_owned()),
-                ("@scope/pkg-foo".to_owned(), "nodeNameScope".to_owned())
-            ])
-        );
-    }
-
-    #[tokio::test]
-    async fn returns_project_using_alias() {
-        let graph = get_aliases_graph(NodeConfig {
-            alias_package_names: Some(NodeProjectAliasFormat::NameAndScope),
-            ..NodeConfig::default()
-        })
-        .await;
-
-        assert_eq!(
-            graph.load("@scope/pkg-foo").unwrap().id,
-            "nodeNameScope".to_owned()
-        );
-    }
-
-    #[tokio::test]
-    async fn graph_uses_id_for_nodes() {
-        let graph = get_aliases_graph(NodeConfig {
-            alias_package_names: Some(NodeProjectAliasFormat::NameAndScope),
-            ..NodeConfig::default()
-        })
-        .await;
-
-        graph.load("pkg-bar").unwrap();
-        graph.load("@scope/pkg-foo").unwrap();
 
         assert_snapshot!(graph.to_dot());
     }
